FPU initialization is added. Fixing makefile to be able use recursivee search.
FPU is now supported by HW initialization. But if there is need to switch it
of, it is possible by removing __FPU_USED definition from Make.cfg.
As second part for FPU support it is added new (math library) which can be used
for complex math operation like fft or fir computaition.

Also it is possible now to create complex structure inside src directory.